requirements.txt
pandas
pillow
selenium
secure-smtplib
oauthlib
requests
msal
requests-oauthlib
pyautogui
keyboard
pyperclip
